feat : add TradeType to Table and Edit Check Logic to filter match tradeType
- Exception Handler에서 ExposedException (Exception) 을 잡을 수 있도록 추가.
- ItemChecker.kt 제거
