+++ src/main/kotlin/taeyun/malanalter/alertitem/domain/AlertItemEntity.kt
@@ -10,6 +10,7 @@ class AlertItemEntity(id: EntityID<Int>) : IntEntity(id) {
     var itemId by AlertItemTable.itemId
     var itemCondition by AlertItemTable.itemCondition
     var isAlarm by AlertItemTable.isAalarm
+    var tradeType by AlertItemTable.tradeType
     var userId by AlertItemTable.userId
     val bids by ItemBidEntity referrersOn ItemBidTable.alertItemId orderBy ItemBidTable.price
 }
