+++ src/main/kotlin/taeyun/malanalter/alertitem/dto/ItemBidInfo.kt
@@ -1,6 +1,5 @@
 package taeyun.malanalter.alertitem.dto
 
-import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonProperty
 
 
@@ -13,21 +12,6 @@ data class ItemBidInfo (
     @JsonProperty("url")
     val url: String,
 ){
-    enum class TradeType {
-        BUY, SELL;
-
-        companion object {
-            @JvmStatic
-            @JsonCreator(mode = JsonCreator.Mode.DELEGATING)
-            fun fromString(value: String): TradeType {
-                return when (value) {
-                    "buy" -> BUY
-                    "sell" -> SELL
-                    else -> throw IllegalArgumentException("Invalid value $value")
-                }
-            }
-        }
-    }
     fun toDiscordMessage(): String {
         return "> 가격=${ItemCondition.changePriceToString(itemPrice)}, comment=${comment}  | [링크](https://mapleland.gg/trade/$url)"
     }
