+++ src/main/kotlin/taeyun/malanalter/alertitem/dto/RegisteredItem.kt
@@ -8,6 +8,7 @@ data class RegisteredItem (
     val itemId: Int,
     val itemOptions: ItemCondition,
     val itemName: String,
+    val tradeType: TradeType,
     val isAlarm: Boolean,
     val userId : Long
 ){
@@ -15,6 +16,7 @@ data class RegisteredItem (
         id = alertItemEntity.id.value,
         itemId = alertItemEntity.itemId,
         itemOptions= alertItemEntity.itemCondition,
+        tradeType = alertItemEntity.tradeType,
         itemName = AlertItemRepository.itemNameMap[alertItemEntity.itemId] ?: "이름 없음",
         isAlarm = alertItemEntity.isAlarm,
         userId = alertItemEntity.userId
