+++ src/main/kotlin/taeyun/malanalter/alertitem/repository/AlertRepository.kt
@@ -4,10 +4,11 @@ import taeyun.malanalter.alertitem.domain.ItemBidEntity
 import taeyun.malanalter.alertitem.dto.ItemBidInfo
 import taeyun.malanalter.alertitem.dto.ItemCondition
 import taeyun.malanalter.alertitem.dto.RegisteredItem
+import taeyun.malanalter.alertitem.dto.TradeType
 
 interface AlertRepository {
     fun getRegisteredItem(): List<RegisteredItem>
-    fun save(itemId: Int, itemCondition: ItemCondition)
+    fun save(itemId: Int, itemCondition: ItemCondition, tradeType: TradeType)
     fun delete(alertId: Int)
     fun update(alertId: Int, updateItemCondition: ItemCondition)
     fun saveItemName(itemId: Int, itemName: String)
